+++ config/test.exs
@@ -1,0 +1,3 @@
+use Mix.Config
+
+config :telnyx, :base_url, "https://localhost:12112/v2"